Refactor: Rename module
